srcs/backend/ai_services/rag_service/requirements.txt
requests
uvicorn
fastapi
slowapi
python-multipart
python-jose
python-dotenv
langchain-ollama
langchain-community
langchain-text-splitters
langchain_chroma
